+++ JDCoordinator/Protocols/JDChildCoordinatorProtocol.swift
@@ -24,23 +24,6 @@ public protocol JDChildCoordinatorProtocol: JDBaseCoordinatorProtocol {
     func setParent(to coordinator: JDParentCoordinatorProtocol)
 }
 
-protocol _JDChildCoordinatorProtocol: JDChildCoordinatorProtocol {
-    var parentCoordinator: JDParentCoordinatorProtocol! { get set }
-}
-
-extension _JDChildCoordinatorProtocol {
-
-    public func setParent(to coordinator: JDParentCoordinatorProtocol) {
-        guard parentCoordinator !== coordinator else {
-            return
-        }
-        guard !coordinator.childCoordinators.contains(self) else {
-            return coordinator.addChild(self)
-        }
-        parentCoordinator = coordinator
-    }
-}
-
 public extension JDChildCoordinatorProtocol {
 
     /// Returns every parentCoordinator.
