+++ JDCoordinator/Types/JDChildCoordinatorType.swift
@@ -1,0 +1,12 @@
+//
+//  JDChildCoordinatorType.swift
+//  JDCoordinator
+//
+//  Created by Jan Dammshäuser on 27.07.17.
+//
+
+import Foundation
+
+public enum JDChildCoordinatorType {
+    case all, except(JDChildCoordinatorProtocol)
+}
